pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "mcp-sdk"
version = "0.1.0"
description = "All-in-one MCP Agent Platform SDK — A2A, A2B, B2B, B2C with rule-based agents"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [
  { name = "MCP SDK Contributors" },
]
keywords = ["mcp", "agent", "a2a", "b2b", "b2c", "llm", "ai", "sdk"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
  # MCP protocol
  "mcp>=1.0.0",
  # Async
  "anyio>=4.5.0",
  "asyncio-mqtt>=0.16.2",
  # HTTP & WebSocket
  "httpx>=0.27.0",
  "websockets>=12.0",
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  # Data validation
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  # Serialization
  "orjson>=3.10.0",
  "PyYAML>=6.0.1",
  "toml>=0.10.2",
  # Auth & Security
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.9",
  # Observability
  "structlog>=24.1.0",
  "opentelemetry-api>=1.25.0",
  "opentelemetry-sdk>=1.25.0",
  # Storage / Memory
  "redis>=5.0.4",
  "sqlalchemy[asyncio]>=2.0.30",
  "aiosqlite>=0.20.0",
  "alembic>=1.13.1",
  # CLI
  "typer[all]>=0.12.3",
  "rich>=13.7.1",
  # gRPC transport (optional)
  "grpcio>=1.64.0",
  "protobuf>=5.27.0",
  # Utilities
  "tenacity>=8.3.0",
  "cachetools>=5.3.3",
  "python-dotenv>=1.0.1",
  "croniter>=2.0.5",
  "jinja2>=3.1.4",
  "black>=26.3.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.7",
  "pytest-cov>=5.0.0",
  "pytest-mock>=3.14.0",
  "httpx>=0.27.0",
  "mypy>=1.10.0",
  "ruff>=0.4.8",
  "black>=24.4.2",
  "pre-commit>=3.7.1",
  "factory-boy>=3.3.0",
]
docs = [
  "mkdocs>=1.6.0",
  "mkdocs-material>=9.5.26",
  "mkdocstrings[python]>=0.25.1",
]
# Optional: geospatial features (H3 spatial indexing)
geo = [
  "h3>=3.7",
  "numpy>=1.26",
]
# Optional: active inference (belief-state math)
inference = [
  "numpy>=1.26",
]
# Optional: exact token counting (OpenAI-compatible)
context = [
  "tiktoken>=0.7",
]
# Install everything
all = [
  "h3>=3.7",
  "numpy>=1.26",
  "tiktoken>=0.7",
]

[project.scripts]
mcp = "mcp_sdk.cli.main:app"

[project.urls]
Homepage = "https://github.com/khulnasoft/mcp-sdk"
Documentation = "https://mcp-sdk.readthedocs.io"
Repository = "https://github.com/khulnasoft/mcp-sdk"
Issues = "https://github.com/khulnasoft/mcp-sdk/issues"

[tool.hatch.build.targets.wheel]
packages = ["mcp_sdk"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "C4", "SIM"]
ignore = ["ANN101", "ANN102"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=mcp_sdk --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pydocstyle]
convention = "google"
add_ignore = ["D100", "D104", "D105", "D107"]

[tool.bandit]
exclude_dirs = ["tests", "docs", "examples"]
skips = ["B101", "B601"]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
version_files = [
  "pyproject.toml:^version",
  "mcp_sdk/__init__.py:^__version__"
]
tag_format = "v$version"
bump_message = "bump: $current_version → $new_version"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "types-pyyaml>=6.0.12.20250915",
]
